+++ src/Modules/BabaPlay.Modules.Associates/Entities/Position.cs
@@ -5,5 +5,4 @@ namespace BabaPlay.Modules.Associates.Entities;
 public class Position : BaseEntity
 {
     public string Name { get; set; } = string.Empty;
-    public int SortOrder { get; set; }
 }